backend/requirements.txt
fastapi
uvicorn[standard]
passlib
python-dotenv
sqlalchemy
pydantic
mysql-connector-python
